Moved Todo specific domain into its own folder
Moved to use MediatR as well.

+++ TodoApi.Api/Todos/TodoRepository.cs
@@ -1,8 +1,7 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using TodoApi.Api.Database;
-using TodoApi.Api.Domain;
 
-namespace TodoApi.Api.Repositories;
+namespace TodoApi.Api.Todos;
 
 public interface ITodoRepository
 {
@@ -10,8 +9,10 @@ public interface ITodoRepository
     Task<Todo> GetTodoById(Guid id);
     Task<Todo> CreateTodoAsync(Todo todo);
     Task<Todo> UpdateTodoAsync(Todo todo);
-    Task DeleteTodoAsync(Guid id);
+    Task<bool> DeleteTodoAsync(Guid id);
     Task<bool> TodoExistsAsync(Guid id);
+    Task<bool> TodoExistsAsync(string task);
+    Task<Todo?> GetToDoByTask(string task);
 }
 
 public class TodoRepository : ITodoRepository
@@ -25,7 +26,7 @@ public class TodoRepository : ITodoRepository
 
     public async Task<IEnumerable<Todo>> GetTodos() => await _context.Todos.ToListAsync();
     public async Task<Todo> GetTodoById(Guid id) => await _context.Todos.FindAsync(id);
-    
+
     public async Task<Todo> CreateTodoAsync(Todo todo)
     {
         await _context.Todos.AddAsync(todo);
@@ -42,12 +43,19 @@ public class TodoRepository : ITodoRepository
         return todo;
     }
 
-    public async Task DeleteTodoAsync(Guid id)
+    public async Task<bool> DeleteTodoAsync(Guid id)
     {
         var todo = await _context.Todos.FindAsync(id);
+        if (todo == null)
+        {
+            return false;
+        }
         _context.Todos.Remove(todo);
-        await _context.SaveChangesAsync();
+        var result = await _context.SaveChangesAsync();
+        return result > 0;
     }
 
     public async Task<bool> TodoExistsAsync(Guid id) => await _context.Todos.AnyAsync(e => e.Id == id);
+    public async Task<bool> TodoExistsAsync(string task) => await _context.Todos.AnyAsync(e => e.Task == task);
+    public async Task<Todo?> GetToDoByTask(string task) => await _context.Todos.FirstOrDefaultAsync(e => e.Task == task);
 }
